requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0
djangorestframework==3.10.3
gunicorn==20.0.4
Pillow==6.2.1
psycopg2==2.8.4
pypng==0.0.20
PyQRCode==1.2.1
pytz==2019.3
sqlparse==0.3.0
whitenoise==4.1.4
